+++ web/src/misc/codeRenderers/codeRenderer.interface.ts
@@ -1,0 +1,10 @@
+/**
+ * Code renderer that handles rendering block of code
+ */
+export interface CodeRenderer
+{
+    /**
+     * Function that handles markdown code and renders correct block code for it
+     */
+    (code: string, language: string|undefined, isEscaped: boolean): string;
+}