+++ web/src/misc/codeRenderers/highlightJsRenderer.ts
@@ -1,0 +1,13 @@
+import hljs from 'highlight.js';
+
+import {CodeRenderer} from './codeRenderer.interface';
+
+/**
+ * Code renderer using highlight js
+ */
+export const HighlightJsCodeRenderer: CodeRenderer = (code: string, language: string|undefined, _isEscaped: boolean) =>
+{
+    const validLanguage = hljs.getLanguage(language) ? language : 'plaintext';
+
+    return `<pre><code class="hljs language-${language}">${hljs.highlight(validLanguage, code).value}</code></pre>`;
+};